requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==7.1.2
comm==0.1.3
cryptography==41.0.2
cssselect==1.2.0
debugpy==1.6.7
decorator==5.1.1
dnspython==2.4.0
ecdsa==0.18.0
email-validator==1.1.2
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.100.0
filelock==3.12.2
fsspec==2023.6.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
huggingface-hub==0.16.4
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.25.0
ipython==8.12.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jupyter_client==8.3.0
jupyter_core==5.3.1
lxml==4.9.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
motor==3.2.0
mpmath==1.2.1
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.4
packaging==23.1
parsel==1.8.1
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==3.9.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.3.0
Pygments==2.15.1
pymongo==4.4.1
pyre-extensions==0.0.29
python-dateutil==2.8.2
python-decouple==3.3
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
pyzmq==25.1.0
regex==2023.6.3
requests==2.31.0
rsa==4.9
safetensors==0.3.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
sympy==1.11.1
tokenizers==0.13.3
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.12.3
uvloop==0.17.0
w3lib==2.1.1
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
zipp==3.16.2
PyJWT
